+++ src/browserless.ts
@@ -25,7 +25,6 @@ import {
   WebSocketRoute,
   WebkitPlaywright,
   availableBrowsers,
-  createLogger,
   getRouteFiles,
   makeExternalURL,
   printLogo,
@@ -48,7 +47,7 @@ type routeInstances =
   | BrowserWebsocketRoute;
 
 export class Browserless extends EventEmitter {
-  protected debug: debug.Debugger = createLogger('index');
+  protected logger: BlessLogger;
   protected browserManager: BrowserManager;
   protected config: Config;
   protected fileSystem: FileSystem;
@@ -96,6 +95,7 @@ export class Browserless extends EventEmitter {
   } = {}) {
     super();
     this.Logger = LoggerOverride ?? BlessLogger;
+    this.logger = new this.Logger('index');
     this.config = config || new Config();
     this.metrics = metrics || new Metrics();
     this.token = token || new Token(this.config);
@@ -139,7 +139,7 @@ export class Browserless extends EventEmitter {
 
     this.metrics.reset();
 
-    this.debug(
+    this.logger.info(
       `Current period usage: ${JSON.stringify({
         date: aggregatedStats.date,
         error: aggregatedStats.error,
@@ -156,7 +156,7 @@ export class Browserless extends EventEmitter {
     );
 
     if (metricsPath) {
-      this.debug(`Saving metrics to "${metricsPath}"`);
+      this.logger.info(`Saving metrics to "${metricsPath}"`);
       this.fileSystem.append(
         metricsPath,
         JSON.stringify(aggregatedStats),
@@ -240,11 +240,18 @@ export class Browserless extends EventEmitter {
     const [[internalHttpRouteFiles, internalWsRouteFiles], installedBrowsers] =
       await Promise.all([getRouteFiles(this.config), availableBrowsers]);
 
+    const hasDebugger = await this.config.hasDebugger();
+    const debuggerURL =
+      hasDebugger &&
+      makeExternalURL(
+        this.config.getExternalAddress(),
+        `/debugger/?token=${this.config.getToken()}`,
+      );
     const docsLink = makeExternalURL(this.config.getExternalAddress(), '/docs');
 
-    this.debug(printLogo(docsLink));
-    this.debug(`Running as user "${userInfo().username}"`);
-    this.debug('Starting import of HTTP Routes');
+    this.logger.info(printLogo(docsLink, debuggerURL));
+    this.logger.info(`Running as user "${userInfo().username}"`);
+    this.logger.info('Starting import of HTTP Routes');
 
     for (const httpRoute of [
       ...this.httpRouteFiles,
@@ -275,12 +282,14 @@ export class Browserless extends EventEmitter {
           this.metrics,
           this.monitoring,
           this.staticSDKDir,
+          this.limiter,
         );
 
         if (!this.routeIsDisabled(route)) {
           route.bodySchema = safeParse(bodySchema);
           route.querySchema = safeParse(querySchema);
           route.config = () => this.config;
+          route.limiter = () => this.limiter;
           route.metrics = () => this.metrics;
           route.monitoring = () => this.monitoring;
           route.fileSystem = () => this.fileSystem;
@@ -291,7 +300,7 @@ export class Browserless extends EventEmitter {
       }
     }
 
-    this.debug('Starting import of WebSocket Routes');
+    this.logger.info('Starting import of WebSocket Routes');
     for (const wsRoute of [
       ...this.webSocketRouteFiles,
       ...internalWsRouteFiles,
@@ -324,11 +333,13 @@ export class Browserless extends EventEmitter {
           this.metrics,
           this.monitoring,
           this.staticSDKDir,
+          this.limiter,
         );
 
         if (!this.routeIsDisabled(route)) {
           route.querySchema = safeParse(querySchema);
           route.config = () => this.config;
+          route.limiter = () => this.limiter;
           route.metrics = () => this.metrics;
           route.monitoring = () => this.monitoring;
           route.fileSystem = () => this.fileSystem;
@@ -359,7 +370,7 @@ export class Browserless extends EventEmitter {
       .map((r) => r.name);
 
     if (duplicateNamedRoutes.length) {
-      this.debug(
+      this.logger.warn(
         `Found duplicate routing names. Route names must be unique:`,
         duplicateNamedRoutes,
       );
@@ -368,7 +379,9 @@ export class Browserless extends EventEmitter {
     httpRoutes.forEach((r) => this.router.registerHTTPRoute(r));
     wsRoutes.forEach((r) => this.router.registerWebSocketRoute(r));
 
-    this.debug(`Imported and validated all route files, starting up server.`);
+    this.logger.info(
+      `Imported and validated all route files, starting up server.`,
+    );
 
     this.server = new HTTPServer(
       this.config,
@@ -381,7 +394,7 @@ export class Browserless extends EventEmitter {
 
     await this.loadPwVersions();
     await this.server.start();
-    this.debug(`Starting metrics collection.`);
+    this.logger.info(`Starting metrics collection.`);
     this.metricsSaveIntervalID = setInterval(
       () => this.saveMetrics(),
       this.metricsSaveInterval,