+++ src/browsers/chromium.playwright.ts
@@ -1,9 +1,9 @@
 import {
   BrowserServerOptions,
   Config,
+  Logger,
   Request,
   ServerError,
-  createLogger,
 } from '@browserless.io/browserless';
 import playwright, { Page } from 'playwright-core';
 import { Duplex } from 'stream';
@@ -15,36 +15,45 @@ export class ChromiumPlaywright extends EventEmitter {
   protected config: Config;
   protected userDataDir: string | null;
   protected running = false;
+  protected logger: Logger;
   protected proxy = httpProxy.createProxyServer();
   protected browser: playwright.BrowserServer | null = null;
   protected browserWSEndpoint: string | null = null;
-  protected debug = createLogger('browsers:chromium:playwright');
   protected executablePath = playwright.chromium.executablePath();
 
   constructor({
     config,
     userDataDir,
+    logger,
   }: {
     config: Config;
+    logger: Logger;
     userDataDir: ChromiumPlaywright['userDataDir'];
   }) {
     super();
 
     this.userDataDir = userDataDir;
     this.config = config;
+    this.logger = logger;
 
-    this.debug(`Starting new browser instance`);
+    this.logger.info(`Starting new ${this.constructor.name} instance`);
   }
 
   protected cleanListeners() {
     this.removeAllListeners();
   }
 
+  public keepAlive() {
+    return false;
+  }
+
   public isRunning = (): boolean => this.running;
 
   public close = async (): Promise<void> => {
     if (this.browser) {
-      this.debug(`Closing browser process and all listeners`);
+      this.logger.info(
+        `Closing ${this.constructor.name} process and all listeners`,
+      );
       this.emit('close');
       this.cleanListeners();
       this.browser.close();
@@ -57,18 +66,22 @@ export class ChromiumPlaywright extends EventEmitter {
   public pages = async (): Promise<[]> => [];
 
   public getPageId = (): string => {
-    throw new ServerError(`#getPageId is not yet supported with this browser.`);
+    throw new ServerError(
+      `#getPageId is not yet supported with ${this.constructor.name}.`,
+    );
   };
 
   public makeLiveURL = (): void => {
     throw new ServerError(
-      `Live URLs are not yet supported with this browser. In the future this will be at "${this.config.getExternalAddress()}"`,
+      `Live URLs are not yet supported with ${this.constructor.name}. In the future this will be at "${this.config.getExternalAddress()}"`,
     );
   };
 
   public newPage = async (): Promise<Page> => {
     if (!this.browser || !this.browserWSEndpoint) {
-      throw new ServerError(`Browser hasn't been launched yet!`);
+      throw new ServerError(
+        `${this.constructor.name} hasn't been launched yet!`,
+      );
     }
     const browser = await playwright.chromium.connect(this.browserWSEndpoint);
     return await browser.newPage();
@@ -78,7 +91,7 @@ export class ChromiumPlaywright extends EventEmitter {
     options: BrowserServerOptions = {},
     version?: string,
   ): Promise<playwright.BrowserServer> => {
-    this.debug(`Launching Playwright Handler`);
+    this.logger.info(`Launching ${this.constructor.name} Handler`);
 
     const opts = {
       ...options,
@@ -95,7 +108,9 @@ export class ChromiumPlaywright extends EventEmitter {
     this.browser = await versionedPw.chromium.launchServer(opts);
     const browserWSEndpoint = this.browser.wsEndpoint();
 
-    this.debug(`Browser is running on ${browserWSEndpoint}`);
+    this.logger.info(
+      `${this.constructor.name} is running on ${browserWSEndpoint}`,
+    );
     this.running = true;
     this.browserWSEndpoint = browserWSEndpoint;
 
@@ -122,7 +137,7 @@ export class ChromiumPlaywright extends EventEmitter {
   };
 
   public proxyPageWebSocket = async () => {
-    console.warn(`Not yet implemented`);
+    this.logger.warn(`${this.constructor.name} Not yet implemented`);
   };
 
   public proxyWebSocket = async (
@@ -138,8 +153,8 @@ export class ChromiumPlaywright extends EventEmitter {
       }
       socket.once('close', resolve);
 
-      this.debug(
-        `Proxying ${req.parsed.href} to browser ${this.browserWSEndpoint}`,
+      this.logger.info(
+        `Proxying ${req.parsed.href} to ${this.constructor.name} ${this.browserWSEndpoint}`,
       );
 
       // Delete headers known to cause issues
@@ -156,7 +171,9 @@ export class ChromiumPlaywright extends EventEmitter {
           target: this.browserWSEndpoint,
         },
         (error) => {
-          this.debug(`Error proxying session: ${error}`);
+          this.logger.error(
+            `Error proxying session to ${this.constructor.name}: ${error}`,
+          );
           this.close();
           return reject(error);
         },